Develop

+++ ThePage/src/ThePage.Core/ViewModels/Author/AuthorDetailViewModel.cs
@@ -30,6 +30,7 @@ namespace ThePage.Core
     {
         readonly IMvxNavigationService _navigation;
         readonly IThePageService _thePageService;
+        readonly IUserInteraction _userInteraction;
 
         #region Properties
 
@@ -83,10 +84,11 @@ namespace ThePage.Core
 
         #region Constructor
 
-        public AuthorDetailViewModel(IMvxNavigationService navigation, IThePageService thePageService)
+        public AuthorDetailViewModel(IMvxNavigationService navigation, IThePageService thePageService, IUserInteraction userInteraction)
         {
             _navigation = navigation;
             _thePageService = thePageService;
+            _userInteraction = userInteraction;
         }
 
         #endregion
@@ -115,6 +117,10 @@ namespace ThePage.Core
             Author.Name = TxtName;
 
             var author = await _thePageService.UpdateAuthor(AuthorBusinessLogic.AuthorCellToAuthor(Author));
+            if (author != null)
+                _userInteraction.ToastMessage("Author updated");
+            else
+                _userInteraction.Alert("Failure updating author");
 
             IsEditing = false;
             IsLoading = false;
@@ -125,14 +131,23 @@ namespace ThePage.Core
             if (IsLoading)
                 return;
 
-            IsLoading = true;
+            if (await _userInteraction.ConfirmAsync("Remove Author?", "Confirm", "DELETE"))
+            {
+                IsLoading = true;
 
-            var result = await _thePageService.DeleteAuthor(AuthorBusinessLogic.AuthorCellToAuthor(Author));
+                var result = await _thePageService.DeleteAuthor(AuthorBusinessLogic.AuthorCellToAuthor(Author));
 
-            if (result)
-                await _navigation.Close(this, true);
-
-            IsLoading = false;
+                if (result)
+                {
+                    _userInteraction.ToastMessage("Author removed");
+                    await _navigation.Close(this, true);
+                }
+                else
+                {
+                    _userInteraction.Alert("Failure removing author");
+                    IsLoading = false;
+                }
+            }
         }
 
         #endregion
