+++ ThePage/src/ThePage.Core/ViewModels/Book/BookDetailViewModel.cs
@@ -30,6 +30,7 @@ namespace ThePage.Core
     {
         readonly IMvxNavigationService _navigation;
         readonly IThePageService _thePageService;
+        readonly IUserInteraction _userInteraction;
 
         #region Properties
 
@@ -110,10 +111,11 @@ namespace ThePage.Core
 
         #region Constructor
 
-        public BookDetailViewModel(IMvxNavigationService navigation, IThePageService thePageService)
+        public BookDetailViewModel(IMvxNavigationService navigation, IThePageService thePageService, IUserInteraction userInteraction)
         {
             _navigation = navigation;
             _thePageService = thePageService;
+            _userInteraction = userInteraction;
         }
 
         #endregion
@@ -152,9 +154,12 @@ namespace ThePage.Core
 
             if (result != null)
             {
+                _userInteraction.ToastMessage("Book updated");
                 Book = BookBusinessLogic.BookToBookCell(result, Authors);
                 SelectedAuthor = Authors.FirstOrDefault(a => a.Id == Book.Author.Id);
             }
+            else
+                _userInteraction.Alert("Failure updating book");
 
             IsEditing = false;
             IsLoading = false;
@@ -165,12 +170,25 @@ namespace ThePage.Core
             if (IsLoading)
                 return;
 
-            IsLoading = true;
+            var answer = await _userInteraction.ConfirmAsync("Remove book?", "Confirm", "DELETE");
+            if (answer)
+            {
+                IsLoading = true;
 
-            var result = await _thePageService.DeleteBook(BookBusinessLogic.BookCellToBook(Book));
+                var result = await _thePageService.DeleteBook(BookBusinessLogic.BookCellToBook(Book));
 
-            if (result)
-                await _navigation.Close(this, true);
+                if (result)
+                {
+                    _userInteraction.ToastMessage("Book removed");
+                    await _navigation.Close(this, true);
+                }
+                else
+                {
+                    _userInteraction.Alert("Failure removing book");
+                    IsLoading = false;
+                }
+
+            }
         }
 
         async Task FetchAuthors()
